Update README.md
The goal of this assignment is to create a simple chat application in Erlang which we dub TinyChat. This document is structured as follows:

Section 1 presents the assignment policies.

Section 2 describes the actors involved in the TinyChat application.

Section 3 describes the message protocol associated to each operation that is to be supported by the actors of TinyChat.

Section 4 presents submission instructions.

Assignment Policies

Collaboration Policy. This homework may be done individually or in pairs. Use of the Internet is allowed, but should not include searching for existing solutions.

Under absolutely no circumstances can code be exchanged between students. Excerpts of code presented in class can be used.

Assignments from previous offerings of the course must not be reused. Violations will be penalized appropriately.

Late Policy. Late submissions are allowed with a penalty of 2 points per hour past the deadline.